charge_amps is not an int, change it to a float

+++ bundles/org.openhab.binding.teslascope/src/main/java/org/openhab/binding/teslascope/internal/api/ChargeState.java
@@ -82,7 +82,7 @@ public class ChargeState {
     public String scheduledChargingStartTime = " ";
 
     @SerializedName("charge_amps")
-    public int chargeAmps;
+    public float chargeAmps;
 
     @SerializedName("charge_current_request")
     public int chargeCurrentRequest;
